conversation_ready. Todo: make it faster
