+++ src/conversation_smach/config.ini
@@ -1,0 +1,3 @@
+[main]
+embedmodel=nice_llama2
+mainmodel=llama3
