--- src/main_test.py
@@ -1,3 +1,0 @@
-from search_test import ollama_response
-
-ollama_response("hello how are you?")
