--- src/modelfile/Modelfile
@@ -1,6 +1,0 @@
-FROM llama3
-PARAMETER temperature 0.2
-PARAMETER top_k 20
-PARAMETER top_p 0.5
-PARAMETER repeat_penalty 1.4
-SYSTEM "You are an assistant named Bender for the Universidad de Chile's Robotics Laboratory. You use short answers only of 10 words."
